card-omap-common: Use signed indices to respect return value.
get_mixer_defaults() was using an unsigned index.  However,
mixer_cache_get_id_by_name() will return an negative error
code.  As a result, program would use an invalid array
index, possibly causing a segfault.  Change type of index
from size_t to int to correct.

+++ card-omap-common-4-5.h
@@ -862,7 +862,7 @@ static struct audio_tool_mixer_cache g_card_mix_defaults = {
 static int get_mixer_defaults(struct audio_tool_mixer_cache *cache)
 {
 	struct audio_tool_mixer_cache *defs = &g_card_mix_defaults;
-	size_t m, n;
+	int m, n;
 	int ret = 0;
 
 	mixer_cache_reset_touch(defs);
